chat-bot-cm-onboard-backend/pyproject.toml
[project]
name = "chat-bot-cm-onboard"
version = "0.1.0"
description = "Chatbot backend using FastAPI + LangChain + OpenAI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "uvicorn>=0.37.0",
    "sqlalchemy>=2.0.43",
    "pymysql>=1.1.2",
    "python-dotenv>=1.1.1",
    "pydantic-settings>=2.11.0",
    "langchain-openai>=0.3.35",
    "openai>=2.2.0",
    "langchain>=0.3.27",
    "langgraph>=0.6.10",
    "ipython>=9.6.0",
    "cryptography>=46.0.3",
]

[tools.uv.scripts]
dev = "uvicorn app.main:app --reload" # run the app with: uv run dev
